push mod fetch details into remote types
