+++ lib/yagu/repos/github/client.go
@@ -2,6 +2,7 @@ package github
 
 import (
 	"context"
+	"fmt"
 	"os"
 
 	"golang.org/x/oauth2"
@@ -15,6 +16,7 @@ func NewClient() (client *github.Client, err error) {
 	ctx := context.Background()
 
 	if token := os.Getenv(TokenEnv); token != "" {
+		fmt.Println("GITHUB OAUTH")
 		ts := oauth2.StaticTokenSource(
 			&oauth2.Token{AccessToken: token},
 		)
